requirements.txt
python_gitlab==3.13.0
GitPython==3.1.0